Skype enhancements
git-svn-id: https://svn.ntop.org/svn/ntop/trunk/nDPI@6439 a02cd0c1-8e76-42e1-a119-56f9641475e2

+++ src/include/ndpi_main.h
@@ -133,7 +133,7 @@ typedef struct node_t {
 /* Least recently used cache */
 
 struct ndpi_LruCacheNumEntry {
-  u_int64_t key;
+  u_int32_t key;
   u_int32_t value;
 };
